requirements.txt
anyio==3.6.2
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
cryptography==38.0.1
dataclasses==0.6
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
gitdb==4.0.9
GitPython==3.1.29
gotrue==0.5.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
invoke==1.7.3
jaraco-classes==3.2.3
jeepney==0.8.0
keyring==23.9.3
more-itertools==9.0.0
packaging==21.3
pip==21.1.3
pkginfo==1.8.3
postgrest-py==0.10.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.10.0
python-semantic-release==7.32.1
readme-renderer==37.2
realtime==0.0.5
requests==2.28.1
requests-toolbelt==0.10.0
rfc3986==1.5.0
SecretStorage==3.3.3
semver==2.13.0
setuptools==56.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.3.5
supabase==0.7.1
supafunc==0.2.2
tomlkit==0.11.5
tqdm==4.64.1
twine==3.8.0
typer==0.6.1
typing-extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
websockets==10.3
wheel==0.37.1
zipp==3.9.0
